let the cache directly inherit from LinkedHashMap
