+++ vertx-sql-client/src/main/java/io/vertx/sqlclient/impl/PreparedStatementCache.java
@@ -1,24 +1,28 @@
 package io.vertx.sqlclient.impl;
 
+import io.vertx.sqlclient.impl.SocketConnectionBase.CachedPreparedStatement;
 import io.vertx.sqlclient.impl.command.CloseStatementCommand;
 
+import java.util.LinkedHashMap;
 import java.util.Map;
 
 /**
- * A LRU replacement strategy based cache for prepared statements.
+ * A LRU replacement strategy cache based on {@link java.util.LinkedHashMap} for prepared statements.
  */
-class PreparedStatementCache extends LruCache<String, SocketConnectionBase.CachedPreparedStatement> {
+class PreparedStatementCache extends LinkedHashMap<String, CachedPreparedStatement> {
+  private final int capacity;
   private final Connection conn;
 
   PreparedStatementCache(int capacity, Connection conn) {
-    super(capacity);
+    super(capacity, 0.75f, true);
+    this.capacity = capacity;
     this.conn = conn;
   }
 
   @Override
-  protected boolean removeEldestEntry(Map.Entry<String, SocketConnectionBase.CachedPreparedStatement> eldest) {
-    boolean needRemove = super.removeEldestEntry(eldest);
-    SocketConnectionBase.CachedPreparedStatement cachedPreparedStatementToRemove = eldest.getValue();
+  protected boolean removeEldestEntry(Map.Entry<String, CachedPreparedStatement> eldest) {
+    boolean needRemove = size() > capacity;
+    CachedPreparedStatement cachedPreparedStatementToRemove = eldest.getValue();
 
     if (needRemove) {
       if (cachedPreparedStatementToRemove.resp.succeeded()) {
@@ -35,11 +39,22 @@ class PreparedStatementCache extends LruCache<String, SocketConnectionBase.Cache
   }
 
   public boolean isReady() {
-    Map.Entry<String, SocketConnectionBase.CachedPreparedStatement> entry = getEldestEntry();
+    Map.Entry<String, CachedPreparedStatement> entry = getEldestEntry();
     if (entry == null) {
       return true;
     } else {
       return entry.getValue().resp != null;
     }
+  }
+
+  public int getCapacity() {
+    return this.capacity;
+  }
+
+  private Map.Entry<String, CachedPreparedStatement> getEldestEntry() {
+    if (size() == 0) {
+      return null;
+    }
+    return (Map.Entry<String, CachedPreparedStatement>) entrySet().toArray()[size() - 1];
   }
 }
